% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros-actions-pkgdown-docs.R
\encoding{UTF-8}
\name{ros_actions_pkgdown_docs}
\alias{ros_actions_pkgdown_docs}
\title{Crea un sitio en \verb{/docs} mediante GitHub Actions}
\usage{
ros_actions_pkgdown_docs(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{overwrite}{Sobrescribe el flujo de trabajo si ya está instalado.}
}
\value{
Devuelve \link[base:NULL]{NULL} de forma invisible. Se llama por sus
efectos secundarios.
}
\description{
Instala un flujo de trabajo de GitHub Actions que genera el sitio
\CRANpkg{pkgdown} del paquete en la carpeta \code{docs} del repositorio.
}
\details{
El resultado final es equivalente a ejecutar \code{\link[=ros_build]{ros_build()}}, pero este comando
se ejecuta en GitHub, no localmente.

Para publicar el sitio en la rama \code{gh-pages} en lugar de la carpeta \code{docs},
usa \code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}}.
}
\section{Generación del sitio}{

Usa \code{\link[=ros_build]{ros_build()}} para generar el sitio de forma local. Consulta
\code{\link[pkgdown:build_site]{pkgdown::build_site()}} para las opciones de generación.
}

\examples{
pkg <- file.path(tempdir(), "pkgdown-docs")
if (!dir.exists(pkg)) {
  dir.create(pkg)
}
ros_actions_pkgdown_docs(pkg)
}
\seealso{
Flujos de trabajo de GitHub Actions:
\code{\link[=ros_actions_check_cron]{ros_actions_check_cron()}},
\code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}}

Sitios pkgdown:
\code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}},
\code{\link[=ros_build]{ros_build()}}
}
\concept{github_actions}
\concept{pkgdown}
