% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros-badge-ropenspain.R
\encoding{UTF-8}
\name{ros_badge_ropenspain}
\alias{ros_badge_ropenspain}
\title{Añade una insignia de \strong{rOpenSpain} a tu archivo \code{README}}
\usage{
ros_badge_ropenspain(install = TRUE)
}
\arguments{
\item{install}{Valor lógico. Si es \code{TRUE}, la insignia se instala en
\code{README.md} o \code{README.Rmd}. Si es \code{FALSE}, muestra un mensaje con el
código Markdown correspondiente.}
}
\value{
Si \code{install} es \code{TRUE}, devuelve un \link[base:logical]{valor lógico} de
forma invisible que indica si se ha añadido la insignia. Si \code{install} es
\code{FALSE}, muestra el código Markdown de la insignia y devuelve
\link[base:NULL]{NULL} de forma invisible. Se llama por sus efectos
secundarios.
}
\description{
Añade una insignia de \strong{rOpenSpain} que redirige a \url{https://ropenspain.es/}
en tu archivo \code{README}:
}
\details{
\ifelse{html}{\href{https://ropenspain.es/}{\figure{ropenspain-badge.svg}{options: alt='rOpenSpain'}}}{**rOpenSpain**}
}
\examples{
ros_badge_ropenspain(install = FALSE)
}
\seealso{
\code{\link[usethis:use_badge]{usethis::use_badge()}} para añadir otras insignias al archivo
\code{README} y \link{ros_pals} para las paletas de colores de \strong{rOpenSpain}.

Recursos de identidad visual:
\code{\link{ros_pals}}
}
\concept{branding}
