% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros-build.R
\encoding{UTF-8}
\name{ros_build}
\alias{ros_build}
\title{Genera tu sitio \CRANpkg{pkgdown} de forma local}
\usage{
ros_build(pkg = ".", ...)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{...}{Argumentos adicionales que se pasan a \code{\link[pkgdown:build_site]{pkgdown::build_site()}}.}
}
\value{
Devuelve \link[base:NULL]{NULL} de forma invisible. Se llama por sus
efectos secundarios.
}
\description{
Esta función envuelve \code{\link[pkgdown:build_site]{pkgdown::build_site()}}.
}
\details{
La única diferencia con \code{\link[pkgdown:build_site]{pkgdown::build_site()}} es que te recuerda los pasos
a seguir para configurar \pkg{rostemplate} y pide confirmación antes de
ejecutar \code{\link[pkgdown:build_site]{pkgdown::build_site()}}. El resultado se genera en la carpeta
\code{docs}.

A diferencia de \code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}}, este proceso se ejecuta de
forma local.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
ros_build()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[pkgdown:build_site]{pkgdown::build_site()}} para las opciones de generación,
\code{\link[=ros_actions_pkgdown_docs]{ros_actions_pkgdown_docs()}} y \code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}} para generar
y publicar el sitio mediante GitHub Actions.

Sitios pkgdown:
\code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}},
\code{\link[=ros_actions_pkgdown_docs]{ros_actions_pkgdown_docs()}}
}
\concept{pkgdown}
