% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_gh_actions_croncheck.R
\name{ros_gh_actions_croncheck}
\alias{ros_gh_actions_croncheck}
\title{Create a GitHub action that checks regularly your package}
\source{
\url{https://github.com/r-lib/actions/blob/master/.github/workflows/check-standard.yaml}
}
\usage{
ros_gh_actions_croncheck(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Path to package.}

\item{overwrite}{Overwrite the action if it was already present.}
}
\description{
The GitHub action created would run R CMD check on your package.
The setup includes a regular check to be run monthly the first day of the
month at 08:30 AM.
}
\details{
Please see \url{https://docs.github.com/en/free-pro-team@latest/actions/reference/events-that-trigger-workflows#schedule}
to learn how cron works.

The default configuration is  \code{'30 08 1 * *'}
("At 08:30 on day-of-month 1.").

Use \href{https://crontab.guru/#30_08_1_*_*}{crontab.guru} to check and
create your own cron tag.
}
