% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_gh_branch.R
\name{ros_gh_branch}
\alias{ros_gh_branch}
\title{Create a github action that deploys on \code{gh-pages} branch}
\usage{
ros_gh_branch(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Path to package.}

\item{overwrite}{Overwrite the action if it was already present.}
}
\description{
The github action created would build a \code{pkgdown} of
your package in the \code{gh-pages} branch of your repo.
}
