% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros-actions-check-cron.R
\encoding{UTF-8}
\name{ros_actions_check_cron}
\alias{ros_actions_check_cron}
\title{Crea un flujo de trabajo de GitHub Actions para comprobar tu paquete}
\usage{
ros_actions_check_cron(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{overwrite}{Sobrescribe el flujo de trabajo si ya está instalado.}
}
\value{
Se llama por sus efectos secundarios y devuelve \code{NULL} de forma
invisible.
}
\description{
El flujo de trabajo de GitHub Actions creado ejecuta \verb{R CMD check} para tu
paquete. La configuración por defecto ejecuta la comprobación el día 1 de
cada mes a las 08:30.
}
\details{
Consulta cómo configurar la ejecución programada en el siguiente enlace:

\url{https://docs.github.com/en/free-pro-team@latest/actions/reference/events-that-trigger-workflows#schedule}

La configuración por defecto es \code{"30 08 1 * *"} ("a las 08:30 del día 1 de
cada mes").

Usa \href{https://crontab.guru/#30_08_1_*_*}{crontab.guru} para crear tu propia
configuración cron.
}
\examples{
pkg <- file.path(tempdir(), "pkgdown-cron")
if (!dir.exists(pkg)) {
  dir.create(pkg)
}
ros_actions_check_cron(pkg)
}
\seealso{
Flujos de trabajo de GitHub Actions:
\code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}},
\code{\link[=ros_actions_pkgdown_docs]{ros_actions_pkgdown_docs()}}
}
\concept{github_actions}
