% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_gh_actions_croncheck.R
\encoding{UTF-8}
\name{ros_actions_check_cron}
\alias{ros_actions_check_cron}
\title{Crea un workflow de GitHub Actions que comprueba tu paquete regularmente}
\usage{
ros_actions_check_cron(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{overwrite}{Sobrescribe el workflow si ya se encuentra instalado.}
}
\value{
Se llama por sus efectos secundarios y devuelve \code{NULL} de forma
invisible.
}
\description{
El workflow de GitHub Actions creado ejecuta \verb{R CMD check} para tu paquete.
La configuración por defecto ejecuta el check de manera mensual los días 1
de cada mes a las 08:30.
}
\details{
Puedes aprender cómo configurar la acción cron en el siguiente enlace:

\url{https://docs.github.com/en/free-pro-team@latest/actions/reference/events-that-trigger-workflows#schedule}

La configuración por defecto es \code{"30 08 1 * *"} ("a las 08:30 del día 1 de
cada mes").

Usa \href{https://crontab.guru/#30_08_1_*_*}{crontab.guru} para crear tu propia
configuración cron.
}
\examples{
pkg <- file.path(tempdir(), "pkgdown-cron")
if (!dir.exists(pkg)) {
  dir.create(pkg)
}
ros_actions_check_cron(pkg)
}
