% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_gh_actions_croncheck.R
\name{ros_actions_check_cron}
\alias{ros_actions_check_cron}
\title{Crea una acción GitHub que comprueba tu paquete de manera regular}
\usage{
ros_actions_check_cron(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{overwrite}{Sobreescribe la acción si ya se encuntra instalada.}
}
\description{
La acción GitHub creada ejecuta \verb{R CMD check} para tu paquete.
La configuración por defecto ejecuta el check de manera mensual los dias 1
de mes a las 08:30 AM.
}
\details{
Puedes aprender cómo configurar la acción cron en el siguiente link:

\url{https://docs.github.com/en/free-pro-team@latest/actions/reference/events-that-trigger-workflows#schedule}

La configuración por defecto  \verb{"30 08 1 * *""} ("A las 08:30 del
dia-del-mes 1.").

Usa \href{https://crontab.guru/#30_08_1_*_*}{crontab.guru} para crear tus propia
configuración cron.
}
