% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_actions_pkgdown_branch.R
\encoding{UTF-8}
\name{ros_actions_pkgdown_branch}
\alias{ros_actions_pkgdown_branch}
\title{Instala un workflow de GitHub Actions que genera tu sitio en \code{gh-pages}}
\usage{
ros_actions_pkgdown_branch(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{overwrite}{Sobrescribe el workflow si ya se encuentra instalado.}
}
\value{
Se llama por sus efectos secundarios y devuelve \code{NULL} de forma
invisible.
}
\description{
Instala un workflow de GitHub Actions que crea el sitio \CRANpkg{pkgdown}
de tu paquete en la rama \code{gh-pages} de tu repositorio.
}
\seealso{
Sitios pkgdown:
\code{\link[=ros_actions_pkgdown_docs]{ros_actions_pkgdown_docs()}},
\code{\link[=ros_build]{ros_build()}}
}
\concept{pkgdown}
