% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros-actions-pkgdown-branch.R
\encoding{UTF-8}
\name{ros_actions_pkgdown_branch}
\alias{ros_actions_pkgdown_branch}
\title{Crea un sitio en \code{gh-pages} mediante GitHub Actions}
\usage{
ros_actions_pkgdown_branch(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{overwrite}{Sobrescribe el flujo de trabajo si ya está instalado.}
}
\value{
Se llama por sus efectos secundarios y devuelve \code{NULL} de forma
invisible.
}
\description{
Instala un flujo de trabajo de GitHub Actions que genera el sitio
\CRANpkg{pkgdown} del paquete en la rama \code{gh-pages} del repositorio.
}
\examples{
pkg <- file.path(tempdir(), "pkgdown-branch")
if (!dir.exists(pkg)) {
  dir.create(pkg)
}
ros_actions_pkgdown_branch(pkg)
}
\seealso{
Sitios pkgdown:
\code{\link[=ros_actions_pkgdown_docs]{ros_actions_pkgdown_docs()}},
\code{\link[=ros_build]{ros_build()}}

Flujos de trabajo de GitHub Actions:
\code{\link[=ros_actions_check_cron]{ros_actions_check_cron()}},
\code{\link[=ros_actions_pkgdown_docs]{ros_actions_pkgdown_docs()}}
}
\concept{github_actions}
\concept{pkgdown}
