% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_actions_pkgdown_branch.R
\name{ros_actions_pkgdown_branch}
\alias{ros_actions_pkgdown_branch}
\title{Instala una acción GitHub que genera tu sitio en la rama \code{gh-pages}}
\usage{
ros_actions_pkgdown_branch(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{overwrite}{Sobreescribe la acción si ya se encuntra instalada.}
}
\description{
Instala una acción GitHub que crea el sitio \CRANpkg{pkgdown} de tu paquete
en la rama \code{gh-pages} de tu repositorio.
}
\seealso{
\code{\link[=ros_build]{ros_build()}}, \code{\link[=ros_actions_pkgdown_docs]{ros_actions_pkgdown_docs()}}
}
