% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_actions_pkgdown_docs.R
\encoding{UTF-8}
\name{ros_actions_pkgdown_docs}
\alias{ros_actions_pkgdown_docs}
\title{Instala un workflow de GitHub Actions que crea tu sitio en \verb{/docs}}
\usage{
ros_actions_pkgdown_docs(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{overwrite}{Sobrescribe el workflow si ya se encuentra instalado.}
}
\value{
Se llama por sus efectos secundarios y devuelve \code{NULL} de forma
invisible.
}
\description{
Este workflow de GitHub Actions genera tu sitio \CRANpkg{pkgdown} en la
carpeta \code{docs} de tu repositorio.
}
\details{
El resultado final es equivalente a ejecutar \code{\link[=ros_build]{ros_build()}}, con la diferencia
de que este comando se ejecuta en GitHub en lugar de ejecutarse localmente.
}
\examples{
pkg <- file.path(tempdir(), "pkgdown-docs")
if (!dir.exists(pkg)) {
  dir.create(pkg)
}
ros_actions_pkgdown_docs(pkg)
}
\seealso{
Sitios pkgdown:
\code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}},
\code{\link[=ros_build]{ros_build()}}
}
\concept{pkgdown}
