% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros-actions-pkgdown-docs.R
\encoding{UTF-8}
\name{ros_actions_pkgdown_docs}
\alias{ros_actions_pkgdown_docs}
\title{Crea un sitio en \verb{/docs} mediante GitHub Actions}
\usage{
ros_actions_pkgdown_docs(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{overwrite}{Sobrescribe el flujo de trabajo si ya está instalado.}
}
\value{
Se llama por sus efectos secundarios y devuelve \code{NULL} de forma
invisible.
}
\description{
Este flujo de trabajo de GitHub Actions genera el sitio \CRANpkg{pkgdown} en
la carpeta \code{docs} del repositorio.
}
\details{
El resultado final es equivalente a ejecutar \code{\link[=ros_build]{ros_build()}}, pero este comando
se ejecuta en GitHub, no localmente.
}
\examples{
pkg <- file.path(tempdir(), "pkgdown-docs")
if (!dir.exists(pkg)) {
  dir.create(pkg)
}
ros_actions_pkgdown_docs(pkg)
}
\seealso{
Sitios pkgdown:
\code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}},
\code{\link[=ros_build]{ros_build()}}

Flujos de trabajo de GitHub Actions:
\code{\link[=ros_actions_check_cron]{ros_actions_check_cron()}},
\code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}}
}
\concept{github_actions}
\concept{pkgdown}
