% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_actions_pkgdown_docs.R
\name{ros_actions_pkgdown_docs}
\alias{ros_actions_pkgdown_docs}
\title{Instala una acción GitHub que crea tu sitio en la carpeta \verb{/docs}}
\usage{
ros_actions_pkgdown_docs(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Ruta a la raíz del paquete.}

\item{overwrite}{Sobreescribe la acción si ya se encuntra instalada.}
}
\description{
Esta acción GitHub genera tu sitio \code{pkgdown} en la carpeta \code{docs} de tu
repositorio.
}
\details{
El resultado final es equivalente a ejecutar \code{\link[=ros_build]{ros_build()}} con la única
diferencia de que este comando se ejecuta en GitHub, en lugar de tener que
ser ejecutado por el usuario.
}
\seealso{
\code{\link[=ros_build]{ros_build()}}, \code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}}
}
