% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_badge_ropenspain.R
\encoding{UTF-8}
\name{ros_badge_ropenspain}
\alias{ros_badge_ropenspain}
\title{Añade una insignia de \strong{rOpenSpain} a tu archivo \code{README}}
\usage{
ros_badge_ropenspain(install = TRUE)
}
\arguments{
\item{install}{Valor lógico. Si es \code{TRUE}, la insignia se instala en
\code{README.md} o \code{README.Rmd}. Si es \code{FALSE}, muestra un mensaje con el
código markdown correspondiente.}
}
\value{
Se llama por sus efectos secundarios. Si \code{install} es \code{FALSE},
muestra el código markdown de la insignia.
}
\description{
Añade una insignia de \strong{rOpenSpain} que redirige a \url{https://ropenspain.es/}
en tu archivo \code{README}:
}
\details{
\ifelse{html}{\href{https://ropenspain.es/}{\figure{ropenspain-badge.svg}{options: alt='rOpenSpain'}}}{**rOpenSpain**}
}
\examples{
ros_badge_ropenspain(install = FALSE)
}
\seealso{
\code{\link[usethis:use_badge]{usethis::use_badge()}}
}
