% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_badge_ropenspain.R
\name{ros_badge_ropenspain}
\alias{ros_badge_ropenspain}
\title{Add a rOpenSpain badge to your \code{README} file}
\usage{
ros_badge_ropenspain(...)
}
\arguments{
\item{...}{ignored}
}
\value{
A markdown badge and a message with the markdown value.
}
\description{
Add an rOpenSpain badge pointing to \url{https://ropenspain.es/} on your \code{README}
file:\ifelse{html}{\href{https://ropenspain.es/}{\figure{ropenspain-badge.svg}{options: alt='rOpenSpain'}}}{\strong{rOpenSpain}}
}
\seealso{
\code{\link[usethis:badges]{usethis::use_badge()}}
}
