% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_build.R
\encoding{UTF-8}
\name{ros_build}
\alias{ros_build}
\title{Genera tu sitio \CRANpkg{pkgdown} en local}
\usage{
ros_build(pkg = ".", ...)
}
\arguments{
\item{pkg}{Path to package.}

\item{...}{
  Arguments passed on to \code{\link[pkgdown:build_site]{pkgdown::build_site}}
  \describe{
    \item{\code{examples}}{Run examples?}
    \item{\code{run_dont_run}}{Run examples that are surrounded in \\dontrun?}
    \item{\code{seed}}{Seed used to initialize random number generation in order to
make article output reproducible. An integer scalar or \code{NULL} for no seed.}
    \item{\code{lazy}}{If \code{TRUE}, will only rebuild articles and reference pages
if the source is newer than the destination.}
    \item{\code{override}}{An optional named list used to temporarily override
values in \verb{_pkgdown.yml}}
    \item{\code{preview}}{If \code{TRUE}, or \code{is.na(preview) && interactive()}, will preview
freshly generated section in browser.}
    \item{\code{devel}}{Use development or deployment process?

If \code{TRUE}, uses lighter-weight process suitable for rapid
iteration; it will run examples and vignettes in the current process,
and will load code with \code{pkgload::load_all()}.

If \code{FALSE}, will first install the package to a temporary library,
and will run all examples and vignettes in a new process.

\code{build_site()} defaults to \code{devel = FALSE} so that you get high fidelity
outputs when you building the complete site; \code{build_reference()},
\code{build_home()} and friends default to \code{devel = TRUE} so that you can
rapidly iterate during development.}
    \item{\code{new_process}}{If \code{TRUE}, will run \code{build_site()} in a separate process.
This enhances reproducibility by ensuring nothing that you have loaded
in the current process affects the build process.}
    \item{\code{install}}{If \code{TRUE}, will install the package in a temporary library
so it is available for vignettes.}
    \item{\code{quiet}}{If \code{FALSE}, generate build messages for build functions that
take \code{quiet} arguments.}
  }}
}
\value{
Se llama por sus efectos secundarios y devuelve \code{NULL} de forma
invisible.
}
\description{
Esta función envuelve \code{\link[pkgdown:build_site]{pkgdown::build_site()}}.
}
\details{
La única diferencia con \code{\link[pkgdown:build_site]{pkgdown::build_site()}} es que te recuerda los pasos
a seguir para configurar \code{rostemplate} y pide confirmación antes de
ejecutar \code{\link[pkgdown:build_site]{pkgdown::build_site()}}. El resultado se genera en la carpeta
\verb{/docs}.

Al contrario que \code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}}, este workflow se ejecuta
localmente.
}
\seealso{
\code{\link[pkgdown:build_site]{pkgdown::build_site()}}.

Sitios pkgdown:
\code{\link[=ros_actions_pkgdown_branch]{ros_actions_pkgdown_branch()}},
\code{\link[=ros_actions_pkgdown_docs]{ros_actions_pkgdown_docs()}}
}
\concept{pkgdown}
