% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_build.R
\name{ros_build}
\alias{ros_build}
\title{Build your pkgdown website}
\usage{
ros_build(pkg = ".", ...)
}
\arguments{
\item{pkg, ...}{See \link[pkgdown]{build_site}.}
}
\description{
A wrapper of \link[pkgdown]{build_site}
}
