% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros-pals.R
\encoding{UTF-8}
\name{ros_pals}
\alias{ros_pals}
\alias{ros_green_pal}
\alias{ros_violet_pal}
\alias{ros_gradient_pal}
\alias{ros_qualitative_pal}
\alias{ros_metro_madrid_pal}
\title{Paletas de colores de \strong{rOpenSpain}}
\usage{
ros_green_pal(n = 4, alpha = 0.9, rev = FALSE)

ros_violet_pal(n = 4, alpha = 0.9, rev = FALSE)

ros_gradient_pal(n = 4, alpha = 0.9, rev = FALSE)

ros_qualitative_pal(n = 6, alpha = 0.9, rev = FALSE)

ros_metro_madrid_pal(n = 4, alpha = 0.9, rev = FALSE)
}
\arguments{
\item{n}{Número de colores.}

\item{alpha}{Valor alfa (transparencia) de los colores. \code{alpha = 1} es opaco
y \code{alpha = 0} es totalmente transparente.}

\item{rev}{Valor lógico. Si es \code{TRUE}, devuelve los colores en orden inverso.}
}
\value{
Un vector de colores.
}
\description{
Paletas de colores basadas en el tema de \strong{rOpenSpain}.
}
\examples{
\dontshow{if (requireNamespace("scales", quietly = TRUE)) withAutoprint(\{ # examplesIf}
scales::show_col(ros_green_pal(9))

scales::show_col(ros_violet_pal(9))

scales::show_col(ros_gradient_pal(9))

scales::show_col(ros_qualitative_pal(9))

scales::show_col(ros_metro_madrid_pal(9))
\dontshow{\}) # examplesIf}
}
\seealso{
Recursos de identidad visual:
\code{\link[=ros_badge_ropenspain]{ros_badge_ropenspain()}}
}
\concept{branding}
\concept{palettes}
