% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros_pals.R
\name{ros_pals}
\alias{ros_pals}
\alias{ros_green_pal}
\alias{ros_violet_pal}
\alias{ros_gradient_pal}
\alias{ros_qualitative_pal}
\title{rOpenSpain paletas de color}
\usage{
ros_green_pal(n = 4, alpha = 0.9, rev = FALSE)

ros_violet_pal(n = 4, alpha = 0.9, rev = FALSE)

ros_gradient_pal(n = 4, alpha = 0.9, rev = FALSE)

ros_qualitative_pal(n = 6, alpha = 0.9, rev = FALSE)
}
\arguments{
\item{n}{Número de colores}

\item{alpha}{Valor alfa (transparencia) de los colores. alpha = 1 es opaco
y alpha = 0 totalmente transparente}

\item{rev}{Valor lógico que indica si los colores se han de mostar en orden
inverso.}
}
\value{
Paletas de colores
}
\description{
Paletas de colores basadas en el tema de rOpenSpain
}
\examples{

scales::show_col(ros_green_pal(9))

scales::show_col(ros_violet_pal(9))

scales::show_col(ros_gradient_pal(9))

scales::show_col(ros_qualitative_pal(9))
}
\concept{ros_plot}
