% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rostemplate-package.R
\docType{package}
\encoding{UTF-8}
\name{rostemplate-package}
\alias{rostemplate}
\alias{rostemplate-package}
\title{rostemplate: Plantilla Para Sitios 'pkgdown' De Paquetes De 'rOpenSpain'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Plantilla privada para crear sitios 'pkgdown' de los paquetes de 'rOpenSpain'. Incluye estilos, recursos estáticos y workflows de 'GitHub Actions' para publicar documentación con una identidad visual coherente.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenspain.github.io/rostemplate/}
  \item \url{https://github.com/ropenspain/rostemplate}
  \item Report bugs at \url{https://github.com/ropenspain/rostemplate/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) (affiliation: rOpenSpain) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) (affiliation: rOpenSpain) [copyright holder]
}

}
\keyword{internal}
