% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rostemplate-package.R
\docType{package}
\name{rostemplate-package}
\alias{rostemplate}
\alias{rostemplate-package}
\title{rostemplate: 'pkgdown' Template for 'rOpenSpain' Packages}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This is a private template for use by core rOpenSpain packages. Please don’t use for your own code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenspain.github.io/rostemplate/}
  \item \url{https://github.com/ropenspain/rostemplate}
  \item Report bugs at \url{https://github.com/ropenspain/rostemplate/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) (rOpenSpain) [copyright holder]

}
\keyword{internal}
